[sweet][Kotlin] Add `Property` component
