+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/objects/PropertyComponent.kt
@@ -1,0 +1,20 @@
+package expo.modules.kotlin.objects
+
+import expo.modules.kotlin.functions.SyncFunctionComponent
+
+data class PropertyComponent(
+  /**
+   * Name of the property.
+   */
+  val name: String,
+
+  /**
+   * Synchronous function that is called when the property is being accessed.
+   */
+  val getter: SyncFunctionComponent? = null,
+
+  /**
+   * Synchronous function that is called when the property is being set.
+   */
+  val setter: SyncFunctionComponent? = null
+)
